+++ src/video_gen/srt_to_images.py
@@ -68,85 +68,3 @@ def generate_images_from_srt(
         )
 
         print("*" * 50)
-
-
-def create_video(
-    image_folder_path: str,
-    audio_file_path: str,
-    save_video_path: str,
-    save_video_path_with_audio: str,
-):
-    # Get a sorted list of all image file paths
-    image_files = sorted(
-        [
-            os.path.join(image_folder_path, img)
-            for img in os.listdir(image_folder_path)
-            if img.endswith(".png")
-        ],
-        key=lambda img: int(img.split("/")[-1].split(".")[0]),
-    )
-
-    # NOTE: TODO: Automatic calculation of the image duration. Use librosa to do it
-    img_duration = 2.16
-
-    # Create a video clip from each image
-    clips = [ImageClip(img).set_duration(img_duration) for img in image_files]
-
-    # Concatenate the clips into a single video clip
-    video = concatenate_videoclips(clips)
-
-    # Write the result to a file
-    video.write_videofile(
-        save_video_path,
-        codec="libx264",
-        fps=32,
-    )
-
-    # Add the audio to the video
-    subprocess.run(
-        [
-            "ffmpeg",
-            "-y",
-            "-i",
-            save_video_path,
-            "-i",
-            audio_file_path,
-            "-c:v",
-            "copy",
-            "-c:a",
-            "aac",
-            "-strict",
-            "experimental",
-            save_video_path_with_audio,
-        ]
-    )
-
-
-"""
-image_folder_path = "results/results_3_5_turbo_16k/"
-audio_file_path = "../data/original/karna_kya_hai_tujhe/karna_kya_hai_tujhe_bpm_111.wav"
-srt_path = "../data/original/karna_kya_hai_tujhe/karna_kya_hai_tujhe_bpm_111_english_translation.srt"
-save_video_path = (
-    "results/video/karna_kya_hai_tujhe_bpm_111_gpt_3_5_16_k_stable_diff.mp4"
-)
-save_video_path_with_audio = "results/video/karna_kya_hai_tujhe_bpm_111_gpt_3_5_16_k_stable_diff_with_context_audio.mp4"
-create_video(
-    image_folder_path,
-    audio_file_path,
-    save_video_path,
-    save_video_path_with_audio,
-)
-"""
-# """
-srt_path = "../data/original/karna_kya_hai_tujhe/karna_kya_hai_tujhe_bpm_111_english_translation.srt"
-context_srt_file_path = "../data/original/karna_kya_hai_tujhe/karna_kya_hai_tujhe_bpm_111_english_translation_context.srt"
-save_images_folder_path = "/Users/neeraj/Work/generative_rap_video/src/results/results_3_5_turbo_16k_w_o_context_base_model/"
-generate_images_from_srt(
-    srt_path,
-    # context_srt_file_path,
-    save_images_folder_path,
-)
-# """
-"""
-ffmpeg -i results/video/karna_kya_hai_tujhe_bpm_111_gpt_3_5_16_k_stable_diff.mp4 -i ../data/original/karna_kya_hai_tujhe/karna_kya_hai_tujhe_bpm_111.wav -c:v copy -c:a aac -strict experimental results/video/karna_kya_hai_tujhe_bpm_111_gpt_3_5_16_k_stable_diff_ffmpeg.mp4
-"""
